No intermediate rootfs.img
Completely replace dmsquash-live with our ostree-live and directly mount
the squash file system without the intermediate rootfs.img

This will result in lower storage needs for the live images,
faster startup and we are in control of the boot process with our
custom dracut module.

See also [1] from ostree-image-creator issues.

[1] https://github.com/lirios/ostree-image-creator/issues/4