Workaround to race condition with mount units
Live image boot will go in emergency mode because sysroot-etc.mount
fails due to a misterious "Protocol error" which seems to be
caused by some race condition, because /sysroot/etc is indeed mounted
even though the unit fails.

The systemd fix for this is not yet being released, so this workaround
is needed to fix the issue.

The same can happen to sysroot-var.mount, so the workaround is for
both units.

See https://github.com/lirios/lirios/issues/69